+++ CHANGELOG.markdown
@@ -1,2 +1,7 @@
+* _v0.8.0_  
+       Move `write*` functions to Core.Program.Logging. Essentially a
+       cosmetic change, but allows us to document the unified output/
+       logging mechanism in one place.
+
 * _v0.7.3_  
-	Initial public release
+       Initial public release